Use existence of build file as first run indicator
If a run fails for any reason, I had to delete or move the whole
Experiment directory or else the script would look for the build
listed in BuildType.txt, which may not exist, causing the script
to break. If I removed the BuildType.txt file, the script would
fail because it didn't exist. This gives us a simple route to fix
up a run that's gotten in a bad state.